+++ packages/core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- CLI: Ignore `--requireReference` if using `--unsafeSkipStorageCheck`. ([#913](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/913))
+- CLI: Throw error if `--requireReference` and `--unsafeSkipStorageCheck` are both enabled. ([#913](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/913))
 
 ## 1.31.0 (2023-10-23)